Add quadratic falloff and adjust example lights.
Added quadratic falloff for point lights. Directional lights do not use
quadratic falloff, as they emulate the sun, which is too far away for
falloff to be perceptible.

Converted the torus and rings to use directional lights, as they are too
big for a simple point light to work properly (too dark! When I
implement power, it will be better.)